Move getVerseTextFromVerseElements from VerseSelection to VerseBoxHelper
